apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0
python-dotenv==1.0.0
httpx==0.27.0
psycopg2-binary==2.9.9
meilisearch==0.31.0
beautifulsoup4==4.12.3
requests==2.32.3
slowapi==0.1.9
itsdangerous==2.2.0
tenacity==9.0.0
lxml==5.3.0
feedparser==6.0.11
python-dateutil==2.9.0
jsonpath-ng==1.6.1
playwright==1.48.0
dateparser==1.2.0
scikit-learn==1.5.0
pdfminer.six==20231228
prometheus-client==0.20.0
statsd==4.0.1
pytesseract==0.3.13
pdf2image==1.17.0
pytest==8.3.0
pytest-asyncio==0.23.0
pytest-cov==5.0.0
